completed tracks state hook to pull a lot of track state related functionality out of huge context file
